Always use DeserializeJson for text/plain /oauth/access_token results
It is possible to setup own DeserializeJson method. It should be used
everywhere, for /oauth/access_token result also - as otherwise
it might be that the client does not expect to receive JsonObject
created by SimpleJson - and wants its own data (which might be just a
simple string, for example, if client has its own JSON parser).