Support user addition to Audience and forcing form data content type in general.
Unit test added for audiences.
VS12 xUnit runner NuGet package dependency added.

Main API change:
Added ForceMultipartFormData to FacebookClient.
Note that this can affect some asyncronous (maybe, as I did not dive into the code)
and multithreaded accesses to the FacebookClient.
You may want to try and expand the signature of the Post methods, I simply did not have the time for it.
Anyhow, this is a backward compatible fix in a sense that it won't affect anyone who does not set this field,
nor those who do but create a new client per request.

You may want to refactor afterwords given the comments above.
As you understand, I needed a quick fix and decided many people may benefit from it even if it may not be perfect.
Thanks for your great work,

Boris