Expose error response and optional transient flag
Some error responses contain the is_transient flag which helps identify
whether the call can be retried, perhaps later. In order to be able to
make this available to the user, add the appropriate property.

Also add the entire error response dictionary so that additional insight
(such as error_data.blame_field_specs) can be inferred by the caller.